+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+5.4.1 / WIP
+==================
+
+Improvements:
+* [OLMIS-5235](https://openlmis.atlassian.net/browse/OLMIS-5235): Improved performance of login:
+** Made the requests when login concurrent (where possible).
+** Minimal Facilities are now cached in the LocalDatabase instead of the LocalStorage.
+
 5.4.0 / 2018-08-16
 ==================